backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv >=0.19.1
sqlalchemy>=1.4.25
passlib>=1.7.0 <= 1.7.4
bcrypt>=3.1.0 <= 3.2.0
python-jose>=3.2.0 <= 3.3.0
python-multipart>=0.0.4 <=  0.0.5
fastapi-mail>=1.0.0 <= 1.0.1

# python-dotenv sqlalchemy passlib bcrypt python-jose python-multipart fastapi-mail mysqlclient